Refactored how opperations are managed and applied.
Now operations can be chained naturally without the creation of another operation.
The output of one will be sent to the other.

There is a distiction for basic operations that do not need parameters.
Normally they apply the operations to all the dice at once.
Now an optional parameter can be give to apply the operation less dice than is rolled.
The resulting output will be array-like.

Updated all the test file and the readme to reflect the changes.