Die Weight values can now be floating point values. Updated the readme and the tests to run cases with floating point values for weights
Also got rid of some magic values related to die value generation.
All the defaults are defined at the top of the file and referenced anywhere they are needed. (even though they all have the same value).
This is because they are used for different things.
This way if they to change them in the future, we can. although I don't see why we would.