first jab at some runtime optimizations.
most of these are making global variable lookups be local.
this will shed off some time since these functions can be called tens of thousands of times.

there are still some things that take a while.
if-else statments currently process all cases due to the implementation.
in order for this to change, a refactor in argument passing needs to be done.